requirements.txt
Django==5.2.8
psycopg2-binary>=2.9.9
djangorestframework==3.16.1
python-dotenv==1.0.0
django-cors-headers==4.3.1
gunicorn==21.2.0
whitenoise==6.6.0
